Create Task

+++ app/src/main/java/com/example/taskmaster/UserProjects.kt
@@ -1,5 +1,6 @@
 package com.example.taskmaster
 
+import androidx.appcompat.app.AlertDialog
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
@@ -82,23 +83,36 @@ class UserProjects : AppCompatActivity(), UserProjectAdapter.UserProjectItemClic
     }
 
     override fun onDeleteUserProject(userProjectId: Int) {
-        apiService.deleteUserProject(userProjectId).enqueue(object : Callback<Void> {
-            override fun onResponse(call: Call<Void>, response: Response<Void>) {
-                if (response.isSuccessful) {
-                    Toast.makeText(this@UserProjects, "User removed from project successfully", Toast.LENGTH_SHORT).show()
-                    fetchProjectTeam()
-                } else {
-                    val errorBody = response.errorBody()?.string()
-                    Log.e("Users", "Failed to remove user from project: ${response.code()} ${response.message()} $errorBody")
-                    Toast.makeText(this@UserProjects, "Failed to remove user from project", Toast.LENGTH_SHORT).show()
-                }
-            }
+        val builder = AlertDialog.Builder(this@UserProjects)
+        builder.setTitle("Confirm Remove User")
+        builder.setMessage("Are you sure you want to remove this user from the project?")
 
-            override fun onFailure(call: Call<Void>, t: Throwable) {
-                Log.e("UserProjects", "Network error: ${t.message}", t)
-                Toast.makeText(this@UserProjects, "Network error", Toast.LENGTH_SHORT).show()
-            }
-        })
+        builder.setPositiveButton("Yes") { dialog, which ->
+            apiService.deleteUserProject(userProjectId).enqueue(object : Callback<Void> {
+                override fun onResponse(call: Call<Void>, response: Response<Void>) {
+                    if (response.isSuccessful) {
+                        Toast.makeText(this@UserProjects, "User removed from project successfully", Toast.LENGTH_SHORT).show()
+                        fetchProjectTeam()
+                    } else {
+                        val errorBody = response.errorBody()?.string()
+                        Log.e("Users", "Failed to remove user from project: ${response.code()} ${response.message()} $errorBody")
+                        Toast.makeText(this@UserProjects, "Failed to remove user from project", Toast.LENGTH_SHORT).show()
+                    }
+                }
+
+                override fun onFailure(call: Call<Void>, t: Throwable) {
+                    Log.e("UserProjects", "Network error: ${t.message}", t)
+                    Toast.makeText(this@UserProjects, "Network error", Toast.LENGTH_SHORT).show()
+                }
+            })
+        }
+
+        builder.setNegativeButton("No") { dialog, which ->
+            dialog.dismiss()
+        }
+
+        val dialog: AlertDialog = builder.create()
+        dialog.show()
     }
 
 }
